backend/requirements.txt
aiosqlite==0.22.1
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.13.0
asyncpg==0.30.0
bcrypt==4.2.0
click==8.3.2
colorama==0.4.6
databases==0.9.0
exceptiongroup==1.3.1
fastapi==0.135.3
greenlet==3.3.2
h11==0.16.0
idna==3.11
passlib==1.7.4
psycopg2-binary==2.9.10
pydantic==2.12.5
pydantic_core==2.41.5
python-jose==3.3.0
python-multipart==0.0.22
SQLAlchemy==2.0.49
starlette==1.0.0
typing-inspection==0.4.2
typing_extensions==4.15.0
uvicorn==0.43.0
